backend/requirements.txt
fastapi
uvicorn
pydantic
openai
qdrant-client
python-dotenv
